+++ src/ui/widgets/__init__.py
@@ -1,0 +1,4 @@
+from .profilerow import ProfileRow
+from .multiline import MultilineEntry
+
+__all__ = ["ProfileRow", "MultilineEntry"]
